Localization, confirmation fix
The numbers on the homepage (reflecting the currently raised amount of
money and amount of backers) are now adjusted to reflect confirmed
donations. This way only if a payment is finalized or a backing is
complete will the backer count increase.

Also added some localization for the confirmation page (not the email
confirmation page, the second page after choosing what you will be
donating).