Hide itemscope div (#23)
Since the video an embed, we do not have applicable divs with relevant
data. As such, the VideoObject referenced is placed in the <head>
and uses the project's name and description (and banner image). In the
previous commit this div would be visible in the DOM, which is of course
not intended behaviour.